Restructure header
When trying to create a different layout for tablets, I found that
placing the author's image in a separate div from the text made it
impossible to wrap the text around the image on smaller screens.

I have restructured the header so have everything be a direct child of it
and used grid to rebuild the initial look instead of relying on flexbox
and an absolutely positioned title. This feels cleaner and more responsive
for the big screen layout as well, so two birds with one stone.